the current db-init.sql is a template it isnt valid without replacement
The postgres image will try to initilize itself with any .sql or
executalble sh files it finds. For circle CI we're using db-init.sql as
template and manually initializing the database, which works, but if you
try to bring up a development environment without updating the template,
it will fail because the .sql file is invalid and the .sh script tries
to do the same thing.

Signed-off-by: Ian Allison <iana@pims.math.ca>